test: Test Google Cloud Endpoint on physical device.
API 17 tablet retrieves joke from joke source library and
displays it.

If there's no Internet connection, or the local computer's
firewall is up causing connection to fail and time out, or
the tablet is not connected to the local computer by USB cable,
the tablet displays default text and an Internet connection error
Toast instead.

The local computer's IP address is used to connect to tablet.
This address is in a resource file called secrets.xml,
which isn't in this repository.